# echango-api-main/swagger-spec.yaml
openapi: 3.0.0
paths:
  /api/v1/requests:
    post:
      operationId: RequestsController_create_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequestDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
      security: &ref_0
        - bearer: []
      tags: &ref_1
        - Requests
    get:
      operationId: RequestsController_findAll_v1
      parameters:
        - name: page
          required: false
          in: query
          schema:
            type: number
        - name: limit
          required: false
          in: query
          schema:
            type: number
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InfinityPaginationRequestResponseDto'
      security: *ref_0
      tags: *ref_1
  /api/v1/requests/{id}:
    get:
      operationId: RequestsController_findById_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
      security: *ref_0
      tags: *ref_1
    patch:
      operationId: RequestsController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
      security: *ref_0
      tags: *ref_1
    delete:
      operationId: RequestsController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
      security: *ref_0
      tags: *ref_1
  /api/v1/requests/{id}/accept:
    patch:
      operationId: RequestsController_accept_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
      security: *ref_0
      tags: *ref_1
  /api/v1/requests/{id}/reject:
    patch:
      operationId: RequestsController_reject_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RejectRequestDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
      security: *ref_0
      tags: *ref_1
  /api/v1/requests/{id}/hold:
    patch:
      operationId: RequestsController_hold_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Request'
      security: *ref_0
      tags: *ref_1
...
  /api/v1/rate-histories:
    post:
      operationId: RateHistoriesController_create_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRateHistoryDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateHistory'
      security: &ref_10
        - bearer: []
      tags: &ref_11
        - Ratehistories
    get:
      operationId: RateHistoriesController_getRateHistoriesByOfficeId_v1
      parameters:
        - name: timePeriod
          required: true
          in: query
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: Get rate histories filtered by time period
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RateHistory'
      security: *ref_10
      tags: *ref_11
  /api/v1/rate-histories/{id}:
    get:
      operationId: RateHistoriesController_findById_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateHistory'
      security: *ref_10
      tags: *ref_11
    patch:
      operationId: RateHistoriesController_update_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRateHistoryDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RateHistory'
      security: *ref_10
      tags: *ref_11
    delete:
      operationId: RateHistoriesController_remove_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
      security: *ref_10
      tags: *ref_11
  /api/v1/working-hours:
    post:
      operationId: WorkingHoursController_create_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkingHourDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkingHour'
      security: &ref_12
        - bearer: []
      tags: &ref_13
        - Workinghours
    get:
      operationId: WorkingHoursController_getMyOfficeWorkingHours_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: Working hours for the authenticated user office
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkingHour'
        '404':
          description: Office not found for user
      security: *ref_12
      summary: Get working hours for authenticated user office
      tags: *ref_13
    patch:
      operationId: WorkingHoursController_updateMyOfficeWorkingHours_v1
      parameters:
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkingHoursDto'
      responses:
        '200':
          description: Working hours updated successfully
        '400':
          description: Invalid data provided
        '404':
          description: Office not found for user
      security: *ref_12
      summary: Update working hours for authenticated user office
      tags: *ref_13
  /api/v1/working-hours/{id}:
    get:
      operationId: WorkingHoursController_findById_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - in: header
          required: false
          name: x-custom-lang
          schema:
            example: en
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkingHour'
      security: *ref_12
      tags: *ref_13
...
info:
  title: ExchangGo24 API
  description: ' API docs'
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    CreateRequestDto:
      type: object
      properties: {}
    FileType:
      type: object
      properties:
        id:
          type: string
          example: cbcfa8b8-3a25-4adb-a9c6-e325f0d0f3ae
        path:
          type: string
          example: https://example.com/path/to/file.jpg
      required:
        - id
        - path
    Role:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
          example: admin
      required:
        - id
        - name
    Status:
      type: object
      properties:
        id:
          type: number
        name:
          type: string
          example: active
      required:
        - id
        - name
    User:
      type: object
      properties:
        id:
          type: number
        email:
          type: string
          example: john.doe@example.com
        provider:
          type: string
          example: email
        socialId:
          type: string
          example: '1234567890'
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        photo:
          $ref: '#/components/schemas/FileType'
        role:
          $ref: '#/components/schemas/Role'
        status:
          $ref: '#/components/schemas/Status'
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        deletedAt:
          format: date-time
          type: string
      required:
        - id
        - email
        - provider
        - socialId
        - firstName
        - lastName
        - photo
        - role
        - status
        - createdAt
        - updatedAt
        - deletedAt
    WorkingHour:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the working hour record
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
        dayOfWeek:
          type: string
          description: Day of the week
          example: Monday
          enum:
            - Monday
            - Tuesday
            - Wednesday
            - Thursday
            - Friday
            - Saturday
            - Sunday
        isActive:
          type: boolean
          description: Indicates if the day is active for working hours
          example: true
        fromTime:
          type: string
          description: Start time of working hours
          example: '09:00'
          nullable: true
        toTime:
          type: string
          description: End time of working hours
          example: '17:00'
          nullable: true
        hasBreak:
          type: boolean
          description: Indicates if there is a break during this day
          example: false
        breakFromTime:
          type: string
          description: Start time of break
          example: '12:00'
          nullable: true
        breakToTime:
          type: string
          description: End time of break
          example: '13:00'
          nullable: true
        officeId:
          type: string
          description: ID of the associated office
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
          nullable: true
        createdAt:
          format: date-time
          type: string
          description: Creation timestamp
          example: '2023-01-01T12:00:00Z'
        updatedAt:
          format: date-time
          type: string
          description: Last update timestamp
          example: '2023-01-02T12:00:00Z'
      required:
        - id
        - dayOfWeek
        - isActive
        - fromTime
        - toTime
        - hasBreak
        - breakFromTime
        - breakToTime
        - officeId
        - createdAt
        - updatedAt
    Office:
      type: object
      properties:
        id:
          type: string
        owner:
          $ref: '#/components/schemas/User'
        officeName:
          type: string
        location:
          type: object
        registrationNumber:
          type: string
        currencyExchangeLicenseNumber:
          type: string
        address:
          type: string
        city:
          type: object
        state:
          type: object
        country:
          type: object
        primaryPhoneNumber:
          type: string
        secondaryPhoneNumber:
          type: object
        thirdPhoneNumber:
          type: object
        whatsappNumber:
          type: object
        logo:
          type: object
        mainImage:
          type: object
        slug:
          type: object
        isActive:
          type: boolean
        isVerified:
          type: boolean
        isFeatured:
          type: boolean
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        deletedAt:
          type: object
        distanceInKm:
          type: number
        rates:
          type: array
          items:
            type: string
        equivalentValue:
          type: number
        images:
          type: object
        workingHours:
          type: array
          items:
            type: string
        email:
          type: object
        todayWorkingHours:
          $ref: '#/components/schemas/WorkingHour'
      required:
        - id
        - owner
        - officeName
        - location
        - registrationNumber
        - currencyExchangeLicenseNumber
        - address
        - city
        - state
        - country
        - primaryPhoneNumber
        - secondaryPhoneNumber
        - thirdPhoneNumber
        - whatsappNumber
        - logo
        - mainImage
        - slug
        - isActive
        - isVerified
        - isFeatured
        - createdAt
        - updatedAt
        - deletedAt
        - distanceInKm
        - rates
        - equivalentValue
        - images
        - workingHours
        - email
        - todayWorkingHours
    Request:
      type: object
      properties:
        id:
          type: string
        office:
          $ref: '#/components/schemas/Office'
        status:
          type: string
        rejectReason:
          type: string
        additionalMessage:
          type: string
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - office
        - status
        - rejectReason
        - additionalMessage
        - createdAt
        - updatedAt
    InfinityPaginationRequestResponseDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Request'
        hasNextPage:
          type: boolean
          example: true
      required:
        - data
        - hasNextPage
    UpdateRequestDto:
      type: object
      properties: {}
    RejectRequestDto:
      type: object
      properties:
        rejectReason:
          type: string
          example: Reject Reason
        additionalMessage:
          type: string
          example: Additional Message
      required:
        - rejectReason
...
    Currency:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        namePlural:
          type: string
        symbol:
          type: string
        symbolNative:
          type: string
        decimalDigits:
          type: number
        rounding:
          type: number
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
        flag:
          type: string
      required:
        - id
        - code
        - name
        - namePlural
        - symbol
        - symbolNative
        - decimalDigits
        - rounding
        - createdAt
        - updatedAt
        - flag
...
    CreateRateHistoryDto:
      type: object
      properties: {}
    RateHistory:
      type: object
      properties:
        id:
          type: string
        office:
          $ref: '#/components/schemas/Office'
        targetCurrency:
          $ref: '#/components/schemas/Currency'
        baseCurrency:
          $ref: '#/components/schemas/Currency'
        oldBuyRate:
          type: number
        oldSellRate:
          type: number
        newBuyRate:
          type: number
        newSellRate:
          type: number
        isActive:
          type: boolean
        createdAt:
          format: date-time
          type: string
        updatedAt:
          format: date-time
          type: string
      required:
        - id
        - office
        - targetCurrency
        - baseCurrency
        - oldBuyRate
        - oldSellRate
        - newBuyRate
        - newSellRate
        - isActive
        - createdAt
        - updatedAt
    UpdateRateHistoryDto:
      type: object
      properties: {}
    CreateWorkingHourDto:
      type: object
      properties:
        dayOfWeek:
          type: string
          description: Day of the week
          enum:
            - Monday
            - Tuesday
            - Wednesday
            - Thursday
            - Friday
            - Saturday
            - Sunday
          example: Monday
        isActive:
          type: boolean
          description: Indicates if the day is active for working hours
          example: true
        fromTime:
          type: string
          description: Start time of working hours (HH:MM format)
          example: '09:00'
        toTime:
          type: string
          description: End time of working hours (HH:MM format)
          example: '17:00'
        hasBreak:
          type: boolean
          description: Indicates if there is a break during this day
          example: false
          default: false
        breakFromTime:
          type: string
          description: Start time of break (HH:MM format)
          example: '12:00'
        breakToTime:
          type: string
          description: End time of break (HH:MM format)
          example: '13:00'
        officeId:
          type: string
          description: ID of the associated office
          example: 3fa85f64-5717-4562-b3fc-2c963f66afa6
      required:
        - dayOfWeek
        - isActive
        - hasBreak
        - officeId
    UpdateWorkingHoursDto:
      type: object
      properties: {}
...